models_init/requirements.txt
APScheduler==3.10.4
certifi==2024.8.30
influxdb-client==1.47.0
numpy==2.0.2
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
reactivex==4.0.4
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
